# go-backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API de acciones y recomendaciones",
        "title": "StockInsight API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api",
    "paths": {
        "/api/recommendations": {
            "get": {
                "description": "Devuelve una lista con 10 acciones recomendadas para comprar, mantener y vender, basadas en la puntuación de los brokers.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Recommendations"
                ],
                "summary": "Recomendaciones de acciones",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/domain.StockRecommendation"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/stocks": {
            "get": {
                "description": "Devuelve acciones con paginación y filtros",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Stocks"
                ],
                "summary": "Lista de acciones",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Número de página",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Cantidad por página",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filtra por ticker (ILIKE)",
                        "name": "ticker",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filtra por nombre de empresa (ILIKE)",
                        "name": "company",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filtra por brokerage (ILIKE)",
                        "name": "brokerage",
                        "in": "query"
                    },
                    {
                        "type": "number",
                        "description": "Filtra por target_from mínimo",
                        "name": "target_from_min",
                        "in": "query"
                    },
                    {
                        "type": "number",
                        "description": "Filtra por target_from máximo",
                        "name": "target_from_max",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Fecha mínima (YYYY-MM-DD)",
                        "name": "date_from",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Fecha máxima (YYYY-MM-DD)",
                        "name": "date_to",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "domain.StockRecommendation": {
            "type": "object",
            "properties": {
                "action": {
                    "type": "string"
                },
                "brokerage": {
                    "type": "string"
                },
                "company": {
                    "type": "string"
                },
                "normalize_rating_from": {
                    "type": "string"
                },
                "normalize_rating_to": {
                    "type": "string"
                },
                "ticker": {
                    "type": "string"
                },
                "weight_score": {
                    "type": "number"
                }
            }
        }
    }
}